lib: Fix app_limited to allow cwnd growth
There is a problem when sender has some leftover data cannot be
sent as a QUIC packet, app_limited is always true so cwnd doesn't
grow well especially during congestion avoidance.

Let's think the following scenario:

- max_packet_size=1200
- at send(): cwnd=1220, flight_size=0, send a full packet, 1200 bytes
- at send(): cwnd=1220, flight_size=1200
   we are only able to send 20 bytes, but it does't fit even for
   QUIC headers (usually larger than 20 bytes including all
   QUIC/HTTP3 headers). send() returns Done.
   Sine flight_size < cwnd, app_limited=true

However this is not true because app already send to up cwnd,
not sending more because of headers. This case should be considered
as cwnd limited.

The fix is to update app_limited when we calculate send packet
size before considering header size. If still limited by cwnd,
app_limited will be false.